Do coverage-driven testing of Buffer.select().
This found one condition that wasn't covered by the existing tests. I also wrote
targeted whitebox tests for each of the interesting cases in the code. Tests are
cheap and confidence is valuable.

One other thing this brought up is that we don't support empty byte strings any
more, which simplifies one case.