Avoid excess ByteArray copies when creating ByteStrings
ByteString.of() was copying the data 2 extra times before passing it to
the commonOf() method. Avoid these extra copies by converting the
commonOf() method to take a ByteArray directly and only copy the array
in that method.

MessageDigests and Macs create unique ByteArrays which can safely be
wrapped directly with a ByteString without having to copy the array.
Avoid the spread operator and using the ByteString.of() create method.