Change BufferedSource.buffer() to be a val in Kotlin.
Same for BufferedSink.buffer().

This is because for Kotlin callers these methods shadow the
Source.buffer() and Sink.buffer() methods that have different
behavior. Methods with different behavior shouldn't have the
same names!

One consequence of this is that it will nudge Java users to
change from BufferedSource.buffer() to BufferedSource.getBuffer()
which is slightly worse. I couldn't find a good workaround due
to the lack of support for @JvmName in Kotlin interfaces.

Most users will never need to use either method.