Create a Source and Sink throughput Throttler
Created a Throttler class which can throttle Sources and Sinks to a
desired throughput. Multiple Sources and Sinks can be attached to the
same Throttler and their combined throughput will not exceed the desired
throughput. Multiple Throttlers can also be used on the same Source or
Sink and they will all be honored.