AsyncTimeout wrapped Source needs to timeout on long close
The wrapping Source for an AsyncTimeout was missing a call to enter() in
close(). This means a timeout would never occur if closing the Source
took longer than then the current timeout.

Added some unit tests to cover this case as well for both Source and
Sink and close() and flush().