Initial native implementation of Filesystem (#793)
I'm working in IntelliJ and it's been happier to track a single
target platform than several. I'll expand this to Linux and iOS
in a follow-up PR.

This satisfies egorand's recommendation of making cwd() a function
rather than a val. Annoyingly the native implementation can fail
in ways far more plausible than the JVM implementation.